Fixed surface reconnection when effect is active
The Google Filter Framework has asynchronous API to stop the effect preview.
The stopping is delayed and disconnect from surface is done in main thread
somewhere later. There is no way to wait for the stop to be finished in
onPause(), because stop callback will be triggered in main thread as well.

In case when activity is quickly paused/resumed between triggering effect
stop and actual stop, surface will be still in connected state and attempt
to give it to Camera device will be aborted with error:

SurfaceTexture: [SurfaceView] connect: already connected (cur=1, req=4)

This patch adds waiting until effect stop will be finished on video
activity resume.

Change-Id: I02198d883235a17b7182ab0c22556ea28bcd8c53
Signed-off-by: Daniel Levin <dendy@ti.com>